+++ package.json
@@ -6,19 +6,11 @@
     "@types/jest": "24.0.21",
     "@types/node": "12.12.3",
     "@types/react": "16.9.11",
-    "@types/react-color": "^3.0.4",
-    "@types/react-dnd-multi-backend": "^6.0.0",
-    "@types/react-dnd-touch-backend": "^0.5.0",
     "@types/react-dom": "16.9.3",
     "babel-loader": "^8.0.6",
     "csstips": "^1.2.0",
     "csx": "^10.0.0",
     "react": "^16.12.0",
-    "react-color": "^2.18.1",
-    "react-dnd": "^11.1.3",
-    "react-dnd-html5-backend": "^11.1.3",
-    "react-dnd-multi-backend": "^6.0.2",
-    "react-dnd-touch-backend": "^11.1.3",
     "react-dom": "^16.11.0",
     "react-redux": "^7.1.3",
     "react-router": "^5.1.2",
