+++ src/LocalStore/Actions.ts
@@ -13,20 +13,13 @@ import { ServerCommunication } from './ServerCommunication';
 import { FieldName } from '../model/Item/FieldName';
 import { ActualIdGenerator } from '../idGenerator/ActualIdGenerator';
 import { Transaction } from '../model/Transaction/Transaction';
-import {
-  hoverItem,
-  draggedItem,
-  toggleFilter,
-  search,
-  order
-} from '../ReduxStore/actions';
 
 const idGen = new ActualIdGenerator();
 
 export class Actions {
   private store: Store;
   constructor(
-    private dispatcher: Dispatcher,
+    dispatcher: Dispatcher,
     localStorage: LocalStorage,
     serverCommunication: ServerCommunication
   ) {
@@ -58,33 +51,6 @@ export class Actions {
     const transaction = new Transaction();
     transaction.add(change);
     this.store.commit(transaction);
-  }
-
-  hover(
-    newHover: {
-      path: string;
-      place: string;
-      id: ItemId;
-      parentId: ItemId | null;
-    } | null
-  ) {
-    this.dispatcher.dispatch(hoverItem(newHover));
-  }
-
-  dragged(itemId: ItemId | null) {
-    this.dispatcher.dispatch(draggedItem(itemId));
-  }
-
-  toggleFilter(filterId: string) {
-    this.dispatcher.dispatch(toggleFilter(filterId));
-  }
-
-  search(searchString: string) {
-    this.dispatcher.dispatch(search(searchString));
-  }
-
-  order(props: { attribute?: string; asc?: boolean }) {
-    this.dispatcher.dispatch(order(props));
   }
 
   createItem(field: FieldName, newValue: any): ItemId {
