+++ src/LocalStore/Store.ts
@@ -1,7 +1,7 @@
 import { ItemId } from '../model/Item/ItemId';
 import { StoredItem } from '../model/Item/StoredItem';
 import { Change, ChangeResponse, ItemChange } from '../model/Change/Change';
-import { ViewItem, HashtagInfo, ResponsibleInfo } from '../model/Item/ViewItem';
+import { ViewItem } from '../model/Item/ViewItem';
 import {
   updateItem,
   createItemList,
@@ -197,14 +197,6 @@ export class Store {
       auxilaryColumns,
       children: this.items[id].getChildren(),
       parents: this.items[id].getParents(),
-      hashtags: this.items[id]
-        .getHashtags()
-        .map(x => (this.items[x] ? this.items[x].getHashtagInfo() : null))
-        .filter(x => !!x) as HashtagInfo[],
-      responsibles: this.items[id]
-        .getResponsibles()
-        .map(x => (this.items[x] ? this.items[x].getResponsibleInfo() : null))
-        .filter(x => !!x) as ResponsibleInfo[],
       updateness: this.items[id].getUpdateness()
     };
   }
