+++ src/model/Item/FieldTypeOf.ts
@@ -1,5 +1,3 @@
-import { getRandomColor } from '../../ReduxStore/commons';
-
 interface FieldType {
   type: string;
   params?: any;
@@ -12,20 +10,12 @@ interface NamedFieldType extends FieldType {
 
 const types: Record<string, FieldType> = {
   title: { type: 'String', getDefaultValue: () => '' },
-  priority: { type: 'Number', getDefaultValue: () => 0 },
   description: { type: 'Text', getDefaultValue: () => '' },
   isIt: { type: 'Boolean', getDefaultValue: () => false },
   deleted: { type: 'Boolean', getDefaultValue: () => false },
   count: { type: 'Number', getDefaultValue: () => 0 },
   due: { type: 'Date', getDefaultValue: () => new Date().toISOString() },
-  color: { type: 'Color', getDefaultValue: () => getRandomColor() },
-  hashtag: { type: 'String', getDefaultValue: () => '' },
-  username: { type: 'String', getDefaultValue: () => '' },
-  responsible: {
-    type: 'Enum',
-    params: { values: ['', 'A', 'Z', '2'] },
-    getDefaultValue: () => ''
-  },
+  color: { type: 'Color', getDefaultValue: () => '#000000' },
   state: {
     type: 'Enum',
     params: { values: ['todo', 'doing', 'done'] },
