+++ src/model/Item/ViewItem.ts
@@ -7,8 +7,6 @@ export interface ViewItem {
   originalFields: Record<string, any>;
   auxilaryColumns: string[];
   children: ItemId[];
-  hashtags: HashtagInfo[];
-  responsibles: ResponsibleInfo[];
   parents: ItemId[];
   updateness: Updateness;
 }
@@ -20,19 +18,4 @@ export interface ViewField {
   params?: any;
   updateness: Updateness;
   auxilaryValues: any[];
-}
-
-export interface ViewItemMeta {
-  viewedChildren: ItemId[];
-}
-
-export interface HashtagInfo {
-  color: string;
-  hashtag: string;
-  id: ItemId;
-}
-
-export interface ResponsibleInfo {
-  username: string;
-  id: ItemId;
 }
