+++ src/model/Relation/RelationType.ts
@@ -1,10 +1,6 @@
 export enum RelationType {
   Child = 'child',
-  Parent = 'parent',
-  Hash = 'hash',
-  HashOf = 'hashof',
-  Responsible = 'responsible',
-  ResponsibleOf = 'responsibleof'
+  Parent = 'parent'
 }
 
 export const oppositeOf = (rel: RelationType): RelationType => {
@@ -13,13 +9,5 @@ export const oppositeOf = (rel: RelationType): RelationType => {
       return RelationType.Parent;
     case RelationType.Parent:
       return RelationType.Child;
-    case RelationType.Hash:
-      return RelationType.HashOf;
-    case RelationType.HashOf:
-      return RelationType.Hash;
-    case RelationType.Responsible:
-      return RelationType.ResponsibleOf;
-    case RelationType.ResponsibleOf:
-      return RelationType.Responsible;
   }
 };
